+++ include/libtorrent/peer_connection.hpp
@@ -458,9 +458,9 @@ namespace libtorrent
 
 		stat const& statistics() const { return m_statistics; }
 		void add_stat(boost::int64_t downloaded, boost::int64_t uploaded);
-		void sent_bytes(int bytes_payload, int bytes_protocol);
-		void received_bytes(int bytes_payload, int bytes_protocol);
-		void trancieve_ip_packet(int bytes, bool ipv6);
+		void sent_bytes(size_t bytes_payload, size_t bytes_protocol);
+		void received_bytes(size_t bytes_payload, size_t bytes_protocol);
+		void trancieve_ip_packet(size_t bytes, bool ipv6);
 		void sent_syn(bool ipv6);
 		void received_synack(bool ipv6);
